ShellRun

Description

This module is the entry point to the ShellRun library used by the ShellRun executable.

Documentation

newtype ShellT e m a

ShellT is the main application type that runs shell commands.

Constructors

MkShellT

Fields

runShellT :: ReaderT e m a

runShell :: (HasCommands env, HasLegend env, MonadReader env m, MonadLogger m, MonadShell m) => m ()

runShell is the entry point for running shell commands, i.e., MonadShell instances.