ShellRun.Async

Description

This module provides the runCommands function used for running a list of commands asynchronously.

Documentation

runCommands :: (HasCommandDisplay env, HasLogQueue env, HasSubLogging env, HasTimeout env, MonadIO m, MonadLogger m, MonadUnliftIO m, MonadReader env m) => [Command] -> m ()

Runs all commands asynchronously while printing out a running counter. When a command finishes/crashes, stdout is updated with the result. If a haskell exception is encountered in shell-run itself, this is considered a fatal error and all threads are killed.