ShellRun.Class.MonadLogger

Description

Provides the MonadLogger class, used for having pretty logs in a monadic setting. Logging via this class includes both textual prefixes (e.g. [INFO], [ERROR]) and also common terminal control prefixes for colors.

Types for logging

class Monad m => MonadLogger m where

MonadLogger is a simple typeclass for abstracting logging functions.

Methods

logNoLine :: Text -> m ()

logLine :: Text -> m ()

data LogLevel

Determines the logging level.

Constructors

Debug

Info

InfoBlue

InfoCyan

InfoSuccess

Warn

Error

Fatal

data LogMode

Determines the logging newline behavior.

Constructors

Line

NoLine

Pretty formatted logging

logLevelMode :: MonadLogger m => LogLevel -> LogMode -> Text -> m ()

This is the most general way to log, includes options for LogLevel and LogMode.

logDebug :: MonadLogger m => Text -> m ()

Debug formatted logLine.

logInfo :: MonadLogger m => Text -> m ()

Info formatted logLine.

logInfoBlue :: MonadLogger m => Text -> m ()

Blue Info formatted logLine.

logInfoCyan :: MonadLogger m => Text -> m ()

Cyan Info formatted logLine.

logInfoSuccess :: MonadLogger m => Text -> m ()

Success Info formatted logLine.

logWarn :: MonadLogger m => Text -> m ()

Warn formatted logLine.

logError :: MonadLogger m => Text -> m ()

Error formatted logLine.

logFatal :: MonadLogger m => Text -> m ()

Fatal formatted logLine.

Functions for manipulating carriage returns

resetCR :: MonadLogger m => m ()

Resets the carriage return.

clearLine :: MonadLogger m => m ()

Clears 80 characters on the given line, then prints a newline.

clearNoLine :: MonadLogger m => m ()

Clears the current line without starting a new one.