ShellRun.Data.Env

Description

Provides core Env types.

"HasX" style typeclasses required for our concrete Env type.

class HasCommandDisplay env where

Retrieves the queue that logs are sent to

Methods

getCommandDisplay :: env -> CommandDisplay

class HasLogQueue env where

Retrieves the queue that logs are sent to

Methods

getLogQueue :: env -> LogQueue

class HasSubLogging env where

Determines if we should log commands' output.

Methods

getSubLogging :: env -> SubLogging

class HasTimeout env where

Timeout, if any.

Methods

getTimeout :: env -> Maybe NonNegative

Types

data Env

The main Env type used by ShellRun. Intended to be used with MonadReader.

Constructors

MkEnv

Fields

legend :: Maybe Text

timeout :: Maybe NonNegative

subLogging :: SubLogging

commandDisplay :: CommandDisplay

logQueue :: LogQueue

commands :: [Text]

data CommandDisplay

Type for determining if we use the command's key for display, rather than the key itself.

Constructors

ShowKey
Display the command's key, if it exists, rather than the key itself.

ShowCommand
Display the command itself, not the key.

data SubLogging

Type for determining if we stream commands' logs.

Constructors

Disabled
No logging of sub-commands

Enabled
Logging of sub-commands