ShellRun.Logging.Log

Description

This module provides the Log type and associated functions.

Types for logging

data Log

Captures the relevant information concerning a specific log (i.e. text, level, and mode).

Constructors

MkLog

Fields

msg :: Text

The Text for a given log.

lvl :: LogLevel

The LogLevel for a given log.

mode :: LogMode

The LogMode for a given log.

data LogMode

Determines the logging newline/carriage return behavior.

Constructors

NewLine

Plain

CarriageReturn

data LogLevel

Determines the logging level.

Constructors

None

SubCommand

Debug

Info

InfoBlue

InfoCyan

InfoSuccess

Warn

Error

Fatal

Helpers for creating logs

logNone :: Text -> Log

Returns Log with msg = txt, lvl = None, mode = NewLine.

logSubCommand :: Text -> Log

Returns Log with msg = txt, lvl = SubCommand, mode = NewLine.

logDebug :: Text -> Log

Returns Log with msg = txt, lvl = Debug, mode = NewLine.

logInfo :: Text -> Log

Returns Log with msg = txt, lvl = Info, mode = NewLine.

logInfoBlue :: Text -> Log

Returns Log with msg = txt, lvl = InfoBlue, mode = NewLine.

logInfoCyan :: Text -> Log

Returns Log with msg = txt, lvl = InfoCyan, mode = NewLine.

logInfoSuccess :: Text -> Log

Returns Log with msg = txt, lvl = InfoSuccess, mode = NewLine.

logWarn :: Text -> Log

Returns Log with msg = txt, lvl = Warn, mode = NewLine.

logError :: Text -> Log

Returns Log with msg = txt, lvl = Error, mode = NewLine.

logFatal :: Text -> Log

Returns Log with msg = txt, lvl = Fatal, mode = NewLine.

Utility functions for associate levels to colors/prefixes.

logToColor :: Log -> Color

Transforms log to a color based on its LogLevel.

logToPrefix :: Log -> Text

Transforms log to a prefix based on its LogLevel.

levelToColor :: LogLevel -> Color

Maps LogLevel to Color.

levelToPrefix :: LogLevel -> Text

Maps LogLevel to 'Prefix'.