ShellRun.Logging.MonadLogger

Description

Provides the MonadLogger class, used for having pretty logs in a monadic setting. Logging via this class includes both textual prefixes (e.g. [INFO], [ERROR]) and also common terminal control prefixes for colors.

Typeclass for logging

class Monad m => MonadLogger m where

MonadLogger is a simple typeclass for abstracting logging functions.

Methods

putLog :: Log -> m ()

clear :: m ()

Pretty formatted logging

putLogNone :: MonadLogger m => Text -> m ()

putLog with logNone.

putLogSubCommand :: MonadLogger m => Text -> m ()

putLog with logSubCommand.

putLogDebug :: MonadLogger m => Text -> m ()

putLog with logDebug.

putLogInfo :: MonadLogger m => Text -> m ()

putLog with logInfo.

putLogInfoBlue :: MonadLogger m => Text -> m ()

putLog with logInfoBlue.

putLogInfoCyan :: MonadLogger m => Text -> m ()

putLog with logInfoCyan.

putLogInfoSuccess :: MonadLogger m => Text -> m ()

putLog with logInfoSuccess.

putLogWarn :: MonadLogger m => Text -> m ()

putLog with logWarn.

putLogError :: MonadLogger m => Text -> m ()

putLog with logError.

putLogFatal :: MonadLogger m => Text -> m ()

putLog with logFatal.