ShellRun.Math.REquals

Description

Provides the REquals class.

Documentation

class REquals a b where

REquals is used for convenience when we would like to compare different types for equality. This makes sense when both types are wrappers over the same type. Because Eq should be favored whenever possible, no instances for REquals a a are provided.

Methods

(=:=) :: a -> b -> Bool