ShellRun.Math.RMult

Description

Provides the RMult typeclass for safe multiplication.

Documentation

class RMult a b where

The RMult typeclass provides multiplication. The intention is to be used with newtype'd numbers that carry some sort of invariant, e.g., NonNegative. This way we can safely add numbers without losing their invariants.

Associated Types

type Prod a b

Methods

(*:*) :: a -> b -> Prod a b infixl 7