ShellRun.Parsing.Commands

Description

Provides functionality for translating Text commands via a LegendMap.

Documentation

translateCommands :: LegendMap -> [Text] -> Either LegendErr [Command]

Returns a list of Text commands, potentially transforming a given string via the LegendMap legend.

Given a command string s, we first check if s exists as a key in legend. If it does not, we return s. If there is a key matching s, i.e.,

legend = fromList [...,(s, v),...]

where \(v = v_1,,\ldots,,v_n\), then we recursively search on each \(v_i\). We stop and return \(v_i\) when it does not exist as a key in the map.

Example:

>>> :{
  let m = Map.fromList
        [ ("cmd1", "one"),
          ("cmd2", "two"),
          ("all", "cmd1,,cmd2,,other")
        ]
      cmds = translateCommands m ["all","blah"]
  in (fmap . fmap) getCommand cmds
:}
Right ["one","two","other","blah"]

Note: If -- when looking up a line -- we detect a cycle, then a CyclicKeyErr will be returned.

>>> :{
  let m = Map.fromList
        [ ("a", "b"),
          ("b", "c"),
          ("c", "a")
        ]
  in translateCommands m ["a"]
:}
Left (CyclicKeyErr "a -> b -> c -> a")